+++ pkg/course/shellSecrets.go
@@ -1,22 +1,21 @@
-package shell
+package course
 
 import (
 	"bytes"
 	"fmt"
 	"os/exec"
 
-	"github.com/pkg/errors"
 	"k8s.io/klog/v2"
 )
 
 // Executor represents a shell script to run
-type Executor struct {
+type executor struct {
 	Executable string
 	Args       []string
 }
 
-// NewExecutor returns a shell.Executor with the given script
-func NewExecutor(script []string) (*Executor, error) {
+// newShellExecutor returns an executor with the given script
+func newShellExecutor(script []string) (*executor, error) {
 	var args []string
 	if len(script) == 1 {
 		args = []string{}
@@ -25,15 +24,16 @@ func NewExecutor(script []string) (*Executor, error) {
 	}
 	path, err := exec.LookPath(script[0])
 	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintf("failed to find executable for ShellExecutor secret: %s", script[0]))
+		return nil, fmt.Errorf("failed to find executable for ShellExecutor secret: %s - %w", script[0], err)
 	}
-	return &Executor{
+	return &executor{
 		Executable: path,
 		Args:       args,
 	}, nil
 }
 
-func (s Executor) Get(key string) (string, error) {
+// Get returns the value of the secret and also satisfies the secrets.Getter interface
+func (s executor) Get(key string) (string, error) {
 	cmd := exec.Command(s.Executable, s.Args...)
 	var stdoutBuf, stderrBuf bytes.Buffer
 
@@ -45,7 +45,7 @@ func (s Executor) Get(key string) (string, error) {
 	if err != nil {
 		klog.V(8).Infof("stdout: %s", outStr)
 		klog.V(7).Infof("stderr: %s", errStr)
-		return "", errors.Wrap(err, fmt.Sprintf("exit code %d running command %s", cmd.ProcessState.ExitCode(), cmd.String()))
+		return "", fmt.Errorf("exit code %d running command %s - %w", cmd.ProcessState.ExitCode(), cmd.String(), err)
 	}
 	return outStr, nil
 }
